requirements.txt
bs4
pymysql
pygsheets
pandas
python-dateutil
requests
PyYAML